Make a single FilterByTable operation.
